upgraded ids to long

+++ VoteVis/src/voteVis/Ballot.java
@@ -4,13 +4,13 @@ import processing.core.*;
 
 public class Ballot {
 	private int[] votes_;
-	private int user_id_;
+	private long user_id_;
 	private PImage user_photo_ = null;
 	private String user_name_;
 	private Gender gender_;
 	private String image_string_;
 	
-	Ballot(int user_id_, String user_name_, Gender gender_, int[] votes_, String image_string_) {
+	Ballot(long user_id_, String user_name_, Gender gender_, int[] votes_, String image_string_) {
 		this.votes_ = votes_;
 		this.user_id_ = user_id_;
 		this.image_string_ = image_string_;
@@ -34,7 +34,7 @@ public class Ballot {
 		return votes_;
 	}
 	
-	int user_id() {
+	long user_id() {
 		return user_id_;
 	}
